+++ node/src/template_chain_spec.rs
@@ -1,7 +1,6 @@
 use crate::chain_spec::*;
 use chain_params::SidechainParams;
 use sc_service::ChainType;
-use sidechain_domain::{MainchainAddressHash, UtxoId};
 use sidechain_runtime::{
 	AuraConfig, BalancesConfig, GrandpaConfig, NativeTokenManagementConfig, RuntimeGenesisConfig,
 	SessionCommitteeManagementConfig, SessionConfig, SidechainConfig, SudoConfig, SystemConfig,
@@ -11,7 +10,7 @@ use sidechain_runtime::{
 /// This code should be run by `partner-chains-cli chain-spec`, to produce JSON chain spec file.
 /// `initial_validators` fields should be updated by the `partner-chains-cli chain-spec`.
 /// Add and modify other fields of `ChainSpec` accordingly to the needs of your chain.
-pub fn chain_spec() -> Result<ChainSpec, EnvVarReadError> {
+pub fn chain_spec() -> Result<ChainSpec, envy::Error> {
 	let runtime_genesis_config = RuntimeGenesisConfig {
 		system: SystemConfig { ..Default::default() },
 		balances: BalancesConfig {
@@ -30,23 +29,17 @@ pub fn chain_spec() -> Result<ChainSpec, EnvVarReadError> {
 			initial_validators: vec![],
 		},
 		sidechain: SidechainConfig {
-			params: SidechainParams {
-				chain_id: from_var("CHAIN_ID")?,
-				governance_authority: from_var::<MainchainAddressHash>("GOVERNANCE_AUTHORITY")?,
-				threshold_numerator: from_var("THRESHOLD_NUMERATOR")?,
-				threshold_denominator: from_var("THRESHOLD_DENOMINATOR")?,
-				genesis_committee_utxo: from_var::<UtxoId>("GENESIS_COMMITTEE_UTXO")?,
-			},
+			params: SidechainParams::read_from_env_with_defaults()?,
 			..Default::default()
 		},
 		polkadot_session_stub_for_grandpa: Default::default(),
 		session_committee_management: SessionCommitteeManagementConfig {
 			// Same as SessionConfig
 			initial_authorities: vec![],
-			main_chain_scripts: read_mainchain_scripts_from_env()?,
+			main_chain_scripts: sp_session_validator_management::MainChainScripts::read_from_env()?,
 		},
 		native_token_management: NativeTokenManagementConfig {
-			main_chain_scripts: read_native_token_main_chain_scripts_from_env()?,
+			main_chain_scripts: sp_native_token_management::MainChainScripts::read_from_env()?,
 			..Default::default()
 		},
 	};